backend/health_app/pyproject.toml
[tool.poetry]
name = "health-backend-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
django-otp = "^1.5.4"
qrcode = "^8.0"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
dj-database-url = "2.1.0"
python-dotenv = "^1.0.1"
whitenoise = "^6.7.0"
gunicorn = "^23.0.0"
cryptography = "^43.0.3"
httpx = "^0.27.2"
django-redis = "^5.4.0"
psycopg2-binary = "^2.9.10"
stellar-sdk = "12.0.0"
pandas = "^2.2.3"
phonenumbers = "^8.13.50"
vonage = "^4.0.0"
channels = {extras = ["daphne"], version = "^4.2.0"}
channels-redis = "^4.2.1"
stripe = "^11.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"